requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==24.1
drf-yasg==1.21.7
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==5.0.0
factory-boy==3.3.0
